Removed the duplicate code
