+++ src/books/post.py
@@ -1,16 +1,7 @@
 import json
 import boto3
 import os
-
-
-# import requests
-def create_dynamodb_client():
-    if os.environ.get("REGION_NAME") == 'localhost':
-        client = boto3.client("dynamodb",region_name=os.environ.get("REGION_NAME"), endpoint_url=os.environ.get("END_POINT"),
-                        aws_access_key_id=os.environ.get("ACESS_KEY"), aws_secret_access_key=os.environ.get("SECRET_KEY"))
-    else:
-        client = boto3.client("dynamodb",region_name=os.environ.get("REGION_NAME"))
-    return client
+from . import create_dynamodb_client
 
 def create_put_item_input():
     return {
@@ -78,7 +69,7 @@ def lambda_handler(event, context):
     print(data)
     return {
             "statusCode": 200,
-            "body": data
+            "body": json.dumps(data)
 
         }
